refactor: define a function to parse the YAML file
